Extract Sidebar and Topbar
