+++ public/devIndex.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-    <title>Super Admin | Orgbyte</title>
+    <title>Super Admin | Orgbyte - DEV</title>
   </head>
   <body>
     <div id="root"></div>
