--- src/layout/LayoutWithTopbarAndSidebar.js
@@ -1,223 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
-import {withStyles} from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ChevronRightIcon from '@material-ui/icons/ChevronRight';
-import Collapse from '@material-ui/core/Collapse';
-import ExpandMore from '@material-ui/icons/ExpandMore';
-import ExpandLess from '@material-ui/icons/ExpandLess';
-import ListSubheader from '@material-ui/core/ListSubheader';
-import DashBoardIcon from '@material-ui/icons/Dashboard';
-import ManageIcon from '@material-ui/icons/Person';
-import AdvanceIcon from '@material-ui/icons/Settings';
-import BackupIcon from '@material-ui/icons/Backup';
-import LogoutIcon from '@material-ui/icons/ExitToApp';
-import {Link} from 'react-router-dom';
-
-const drawerWidth = 240;
-
-const styles = theme => ({
-  root: {
-    flexGrow: 1,
-    height: '100%',
-    zIndex: 1,
-    overflow: 'hidden',
-    position: 'relative',
-    display: 'flex'
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    }),
-    backgroundColor: '#5E1619'
-  },
-  appBarShift: {
-    marginLeft: drawerWidth,
-    width: `calc(100% - ${drawerWidth}px)`,
-    transition: theme.transitions.create(['width', 'margin'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen
-    })
-  },
-  menuButton: {
-    marginLeft: 12,
-    marginRight: 36
-  },
-  hide: {
-    display: 'none'
-  },
-  drawerPaper: {
-    position: 'relative',
-    whiteSpace: 'nowrap',
-    width: drawerWidth,
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen
-    }),
-    backgroundColor: '#5E1619'
-  },
-  drawerPaperClose: {
-    overflowX: 'hidden',
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen
-    }),
-    width: theme.spacing.unit * 7,
-    [theme.breakpoints.up('sm')]: {
-      width: theme.spacing.unit * 9
-    }
-  },
-  toolbar: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    padding: '0 8px',
-    ...theme.mixins.toolbar
-  },
-  content: {
-    flexGrow: 1,
-    backgroundColor: theme.palette.background.default,
-    padding: theme.spacing.unit * 3
-  },
-  nested: {
-    paddingLeft: theme.spacing.unit * 4
-  },
-  listHeader: {
-    fontSize: '10px'
-  }
-});
-
-class LayoutWithTopbarAndSidebar extends React.Component {
-  state = {
-    open: false,
-    nestedOpen: false
-  };
-
-  handleDrawerOpen = () => {
-    this.setState({open: true});
-  };
-
-  handleDrawerClose = () => {
-    this.setState({open: false});
-  };
-
-  handleClick = () => {
-    this.setState(state => ({nestedOpen: !state.nestedOpen}));
-  };
-
-  render() {
-    const {classes, theme, children} = this.props;
-
-    return (
-      <div className={classes.root}>
-        <AppBar
-          position="absolute"
-          className={classNames(classes.appBar, this.state.open && classes.appBarShift)}
-        >
-          <Toolbar disableGutters={!this.state.open}>
-            <IconButton
-              color="inherit"
-              aria-label="open drawer"
-              onClick={this.handleDrawerOpen}
-              className={classNames(classes.menuButton, this.state.open && classes.hide)}
-            >
-              <MenuIcon />
-            </IconButton>
-            <Typography variant="title" color="inherit" noWrap>
-              SUPER ADMIN
-            </Typography>
-          </Toolbar>
-        </AppBar>
-        <Drawer
-          variant="permanent"
-          classes={{
-            paper: classNames(classes.drawerPaper, !this.state.open && classes.drawerPaperClose)
-          }}
-          open={this.state.open}
-        >
-          <div className={classes.toolbar}>
-            <IconButton onClick={this.handleDrawerClose}>
-              {theme.direction === 'rtl' ? <ChevronRightIcon /> : <ChevronLeftIcon />}
-            </IconButton>
-          </div>
-          <Divider />
-          <List
-            component="nav"
-            subheader={<ListSubheader className={!this.state.open && classes.hide} component="div">Administration</ListSubheader>}
-          >
-            <Link to="/events" style={{textDecoration: 'none'}}>
-              <ListItem button>
-                <ListItemIcon>
-                  <DashBoardIcon />
-                </ListItemIcon>
-                <ListItemText primary="Dashboard" />
-              </ListItem>
-            </Link>
-
-            <ListItem button>
-              <ListItemIcon>
-                <ManageIcon />
-              </ListItemIcon>
-              <ListItemText primary="Manage Accounts" />
-            </ListItem>
-
-            <ListItem button onClick={this.handleClick}>
-              <ListItemIcon>
-                <AdvanceIcon />
-              </ListItemIcon>
-              <ListItemText inset primary="Advanced" />
-              {this.state.nestedOpen ? <ExpandLess /> : <ExpandMore />}
-            </ListItem>
-            <Collapse in={this.state.nestedOpen} timeout="auto" unmountOnExit>
-              <List component="div" disablePadding>
-                <ListItem button className={classes.nested}>
-                  <ListItemIcon>
-                    <BackupIcon />
-                  </ListItemIcon>
-                  <ListItemText inset primary="Backup" />
-                </ListItem>
-              </List>
-            </Collapse>
-          </List>
-          <List
-            component="nav"
-            subheader={<ListSubheader className={!this.state.open && classes.hide} component="div">User</ListSubheader>}
-          >
-            <ListItem button>
-              <ListItemIcon>
-                <LogoutIcon />
-              </ListItemIcon>
-              <ListItemText primary="Logout" />
-            </ListItem>
-          </List>
-        </Drawer>
-        <main className={classes.content}>
-          <div className={classes.toolbar} />
-          {children}
-        </main>
-      </div>
-    );
-  }
-}
-
-LayoutWithTopbarAndSidebar.propTypes = {
-  children: PropTypes.node.isRequired,
-  classes: PropTypes.object.isRequired,
-  theme: PropTypes.object.isRequired
-};
-
-export default withStyles(styles, {withTheme: true})(LayoutWithTopbarAndSidebar);
